# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para la plataforma LogiRed de transporte de paquetes",
        "title": "LogiRed API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/cars": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Obtener los vehículos del conductor autenticado",
                "responses": {
                    "200": {
                        "description": "lista de carros",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "no autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Un conductor registra su carro con imágenes",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "Crear un nuevo vehículo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Placa",
                        "name": "car_registration",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Marca",
                        "name": "brand",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Modelo",
                        "name": "model",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Color",
                        "name": "color",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Capacidad máxima (kg)",
                        "name": "max_capacity",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Imagen frontal",
                        "name": "frontview_image",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Imagen trasera",
                        "name": "backview_image",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Imagen de placas",
                        "name": "plates_image",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Imagen del espacio",
                        "name": "space_image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "mensaje y carro creado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error en los datos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "no autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/devices/token": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Asocia un token FCM y nombre de dispositivo al usuario autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Registrar dispositivo para notificaciones",
                "parameters": [
                    {
                        "description": "Datos del dispositivo",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "mensaje de éxito",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error en la solicitud",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "no autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/drivers/city/{city}": {
            "get": {
                "description": "Devuelve una lista de conductores que trabajan en una ciudad (búsqueda por LIKE)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Obtener conductores por ciudad",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre de la ciudad (p.ej. 'Tuxtla')",
                        "name": "city",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lista de conductores",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ciudad no especificada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/proposals/{id}/accept": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cambia el estado de una propuesta (1 = aceptada, 3 = rechazada)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proposals"
                ],
                "summary": "Aceptar o rechazar una propuesta",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la propuesta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo estado",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "mensaje de éxito",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error en la solicitud",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "no autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Escribe \"Bearer \u003ctoken\u003e\" para autenticarte",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}